problems/problem-1/app/database.py
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker, declarative_base
from sqlalchemy.exc import OperationalError
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")

engine = create_engine(DATABASE_URL)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = declarative_base()


# Quick test function
def test_connection():
    try:
        with engine.connect() as conn:
            print("✅ Database connection successful!")
    except OperationalError as e:
        print("❌ Database connection failed:", e)

