problems/problem-1/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
python-multipart